+++ static-src/script/app/sudoku/cell.js
