+++ static-src/script/app/sudoku/set.js
