+++ static-src/script/app/sudoku/settings.js
