+++ static-src/script/app/sudoku/utils.js
